PyTTEP/ref/sample-training-code/requirements.txt
absl-py==2.1.0
filelock==3.16.1
fsspec==2024.10.0
grpcio==1.67.1
importlib_metadata==8.5.0
intel-openmp==2021.4.0
Jinja2==3.1.4
Markdown==3.7
MarkupSafe==3.0.2
mkl==2021.4.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.24.1
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.0.0
protobuf==5.28.3
python-dateutil==2.9.0.post0
pytz==2024.2
scipy==1.13.1
six==1.16.0
sympy==1.13.3
tbb==2021.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
torch==2.3.0
torchvision==0.18.0
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.1.3
zipp==3.21.0
